🐛 Fix: Only alert when message can't be deleted, instead of not found

+++ handler/message.mjs
@@ -63,16 +63,16 @@ export async function deleteMessage(msg, alertOnFailure, delay) {
     try {
         if (delay)
             setTimeout(() => {
-                bot.telegram.deleteMessage(msg.chat.id, msg.message_id)
+                bot.tg.deleteMessage(msg.chat.id, msg.message_id)
                     .catch(() => {
                     });
             }, delay);
         else
-            await bot.telegram.deleteMessage(msg.chat.id, msg.message_id);
+            await bot.tg.deleteMessage(msg.chat.id, msg.message_id);
         log(`Chat ${msg.chat.id}: 尝试删除消息，ID: ${msg.message_id}` + (delay ? `，延迟 ${delay} 毫秒` : ''));
     } catch (e) {
-        if (alertOnFailure) {
-            let delMsg = await bot.telegram.sendMessage(msg.chat.id, strings.deleteMsgFailure);
+        if (alertOnFailure && e.message === "can't be deleted") {
+            let delMsg = await bot.tg.sendMessage(msg.chat.id, strings.deleteMsgFailure);
             await deleteMessage(delMsg, false, 15000);
         }
         log(`Chat ${msg.chat.id}: 尝试删除消息失败，ID: ${msg.message_id}，原因：${e.message}`);
@@ -99,7 +99,7 @@ export async function getQueryChatId(ctx) {
         }
     }
     try {
-        let result = await ctx.telegram.getChat(queryChatId);
+        let result = await ctx.tg.getChat(queryChatId);
         if (result.type === 'channel')
             return [result.id.toString(), result.title];
         else {
